+++ firebolt_dest/factory.py
@@ -141,7 +141,9 @@ class firebolt(Destination[FireboltClientConfiguration, "FireboltClient"]):
     def __init__(
         self,
         credentials: Union[FireboltCredentials, Dict[str, Any], str] = None,
+        *,
         s3_location_name: str = "firebolt_s3",
+        s3_location_url: str = "",
         s3_prefix: str = "dlt-landing",
         staging_mode: str = "upload",
         use_schema_per_dataset: bool = False,
@@ -152,6 +154,7 @@ class firebolt(Destination[FireboltClientConfiguration, "FireboltClient"]):
         super().__init__(
             credentials=credentials,
             s3_location_name=s3_location_name,
+            s3_location_url=s3_location_url,
             s3_prefix=s3_prefix,
             staging_mode=staging_mode,
             use_schema_per_dataset=use_schema_per_dataset,